requirements.txt
# Core dependencies for IRIS RAG system
psutil>=5.9.0
GPUtil>=1.4.0
requests>=2.28.0
PyYAML>=6.0

# Document processing
pdfplumber>=0.11.0
sentence-transformers>=2.2.0

# Vector database
chromadb>=0.4.0
numpy>=1.24.0

# LLM inference
ollama>=0.4.0

# Web interface
flask>=2.3.0